# Small performance improvement of PathParser